% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_city.R
\name{add_city}
\alias{add_city}
\title{Add Takoma Park city boundaries
Function to add Takoma Park city boundaries to a leaflet map}
\usage{
add_city(basemap)
}
\arguments{
\item{basemap}{A leaflet map.}
}
\value{

}
\description{
Add Takoma Park city boundaries
Function to add Takoma Park city boundaries to a leaflet map
}
