% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wards.R
\name{add_wards}
\alias{add_wards}
\title{Add Takoma Park ward boundaries}
\usage{
add_wards(basemap)
}
\arguments{
\item{basemap}{A leaflet map.}
}
\value{

}
\description{
Adds polygons for Takoma Park's ward boundaries to the map, with labels
}
\examples{
{
leaflet::leaflet() \%>\%
leaflet::addTiles() \%>\%
add_wards()
}
}
