% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wards.R
\name{add_wards_new}
\alias{add_wards_new}
\title{Add Takoma Park Ward boundaries from 2022 on}
\usage{
add_wards_new(
  basemap,
  .inclabs = T,
  .weight = 1.5,
  .color = "black",
  .opacity = 1
)
}
\arguments{
\item{basemap}{Basemap to add boundaries to}

\item{.inclabs}{Whether to include Ward labels in center of Ward; default T}

\item{.weight}{Weight of Ward boundary lines; deafult 1.5}

\item{.color}{Color of Ward boundary lines; default "black"}

\item{.opacity}{Opacity of Ward boundary lines; default 1}
}
\value{

}
\description{
Adds outlines of Takoma Park's ward boundaries from 2022 on to a leaflet map
}
