% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpoly_legend.R
\name{addbounds}
\alias{addbounds}
\title{Add polygon boundaries}
\usage{
addbounds(
  leafobj,
  placetype,
  spatialdf,
  .color = "black",
  .weight = 0.5,
  .opacity = 0.5,
  labformula = NULL
)
}
\arguments{
\item{leafobj}{Leaflet map adding boundaries layer to.}

\item{placetype}{Group-name as string.}

\item{spatialdf}{Spatial dataframe showing boundaries for}

\item{.color}{Color of polygon boundaries; default "black"}

\item{.weight}{Weight of polygon boundaries; default 0.5}

\item{.opacity}{Opacity of polygon boundaries; default 0.5}

\item{labformula}{Label parameter in addPolygons(); e.g., ~ NAME, or leafletwrappers::label_output() results. Default NULL, so no label displayed}
}
\value{
Leaflet map with layer for polygon boundaries added.
}
\description{
Quick function for adding styled polygon boundaries layer to a leaflet map, with no fill-in.
}
