% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpoly_legend.R
\name{addcluster_legend}
\alias{addcluster_legend}
\title{Add cluster layer and legend}
\usage{
addcluster_legend(
  leafletobj,
  var,
  pal_funct,
  grp,
  labvar,
  popuptext,
  title = grp,
  radius = 0.1,
  weight = 20,
  df = leaflet::getMapData(leafletobj),
  ...
)
}
\arguments{
\item{leafletobj}{Leaflet object}

\item{var}{String name of variable colored in cluster layer}

\item{pal_funct}{Palette function to color cluster layer}

\item{grp}{Group of cluster layer}

\item{labvar}{Label to show when hover over a point; can feed either a data-masked variable name, or a set of labels created by leafletwrappers::label_standard function}

\item{popuptext}{Labels to show in pop-up over point, created by leafletwrappers::label_output function}

\item{title}{Title of legend; default grp}

\item{radius}{Radius of points; default 0.1}

\item{weight}{Weight of points; default 20}

\item{df}{Spatial dataframe behind layer; default leaflet::getMapData(leafletobj)}

\item{...}{Additional .label or .popup arguments for leafletwrappers::addcluster_standard, or opacity argument for leafletwrapper::addlegend_standard}
}
\value{
Leaflet object with cluster layer and legend
}
\description{
Adds cluster layer and legend to leaflet map, with points in cluster layer shaded by palette function
}
