% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpoly_legend.R
\name{addcluster_standard}
\alias{addcluster_standard}
\title{Add standard clustered-points}
\usage{
addcluster_standard(
  leafletobj,
  var,
  pal_funct,
  grp,
  labvar,
  popuptext,
  radius = 0.1,
  weight = 20,
  df = leaflet::getMapData(leafletobj),
  .popup_maxheight = 200,
  .popup_maxwidth = 200,
  .label_padding = "0.2px 0.2px",
  .label_lineheight = 0.8,
  .label_textsize = "10px",
  .label_opacity = 0.8
)
}
\arguments{
\item{leafletobj}{Leaflet object}

\item{var}{Variable layer is being colored by as character string}

\item{pal_funct}{Palette function to color variable by, created by one of leaflet's color functions}

\item{grp}{Group layer belongs to}

\item{labvar}{Label to show when hover over a point; can feed either a data-masked variable name, or a set of labels created by leafletwrappers::label_standard function}

\item{popuptext}{Labels to show in pop-up over point, created by leafletwrappers::label_output function}

\item{radius}{Radius of points; default 0.1}

\item{weight}{Weight of points; default 20}

\item{df}{Spatial dataframe behind layer; default leaflet::getMapData(leafletobj)}

\item{.popup_maxheight}{Maximum height of pop-up window; default 200}

\item{.popup_maxwidth}{Maximum width of pop-up window; default 200}

\item{.label_padding}{Label padding of label; default}

\item{.label_lineheight}{Label line-height; default 0.8}

\item{.label_textsize}{Label textsize; default "10px"}

\item{.label_opacity}{Label opacity; default 0.8}
}
\description{
Standard function for adding clustered points layer to leaflet object, colored by a palette function
}
