% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpoly_legend.R
\name{addpoly_legend}
\alias{addpoly_legend}
\title{Add standard polygon/legend
A function to add a standardized legend/polygon to a leaflet map}
\usage{
addpoly_legend(
  basemap_select,
  df_select,
  pal_funct_select,
  variable_select,
  group_select,
  title_select,
  labels_select
)
}
\arguments{
\item{basemap_select}{A leaflet basemap}

\item{df_select}{A spatial dataframe with the geography stripped}

\item{pal_funct_select}{A color function}

\item{variable_select}{A variable to color in the polygon from the spatial datafrmae}

\item{group_select}{A group for the legend/polygon}

\item{title_select}{The title of the legend/polygon}

\item{labels_select}{A list of labels corresponding to the length of the dataframe to appear as pop-ups}
}
\value{

}
\description{
Add standard polygon/legend
A function to add a standardized legend/polygon to a leaflet map
}
