% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpoly_standard.R
\name{addpoly_standard}
\alias{addpoly_standard}
\title{Add Polygons
A function to add a filled-in polygon layer coloring the polygon by the provided variable}
\usage{
addpoly_standard(basemap, df, pal_funct, variable, group, labels)
}
\arguments{
\item{basemap}{A leaflet basemap}

\item{df}{A spatial dataframe with the geography stripped that provides data visualized in the polygon}

\item{pal_funct}{A color function; e.g. leafletwrappers::color_numeric will often be approriate for numeric functions}

\item{variable}{The variable to visualize in the polygon from the provided dataframe}

\item{group}{The group-name of the polygon}

\item{labels}{A list of labels corresponding to the length of the dataframe to appear as pop-ups}
}
\value{
A polygon layer in a leaflet map
}
\description{
Add Polygons
A function to add a filled-in polygon layer coloring the polygon by the provided variable
}
