% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_standardize.R
\name{addtext}
\alias{addtext}
\title{Add text to leaflet map}
\usage{
addtext(leafletobj, text, pos = "bottomleft")
}
\arguments{
\item{leafletobj}{Leaflet map object}

\item{text}{Text that's displayed on the leaflet map}

\item{pos}{Position of the text; default "bottomleft"}
}
\value{
Leaflet map with displayed text
}
\description{
Add text to a leaflet map; e.g., instructions for how to use the map
}
