% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_nhgis_vars.R
\name{bg_rename}
\alias{bg_rename}
\title{Rename NHGIS variables
Function to rename NHGIS variables to more descriptive names. Intended use is in a rename_all function, with the variable name fed as the string.}
\usage{
bg_rename(string, year_val = 2019)
}
\arguments{
\item{string}{NHGIS variable name.}

\item{year_val}{The selected year of NHGIS data; defaults 2019 for ACS 2015-2019.}
}
\value{
More descriptive NHGIS variable name
}
\description{
Rename NHGIS variables
Function to rename NHGIS variables to more descriptive names. Intended use is in a rename_all function, with the variable name fed as the string.
}
