% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city}
\alias{city}
\title{City Boundaries}
\format{
A spatial dataframe with 6 rows and 8 variables
\describe{
\item{PERIMETER}{Of City}
\item{AREA_NAME}{Name of City}
\item{INCORPORAT}{Whether incorporated}
\item{ACRES}{Acres of City}
\item{OFFICIAL_N}{Official name of City}
\item{ADDRESS}{Address of City offices}
\item{CITY}{City name}
\item{ZIP_CODE}{Zip-code associated with City (note; Takoma Park covers part of the 20912 and 20911 zip codes)}
\item{WEBSITE}{City website address}
\item{MML_MEMBER}{Whether City is part of Maryland Municipal League}
\item{CODE}{Code for City}
\item{SHAPE_Length}{Length of polygon}
\item{SHAPE}{Geometry}
}
}
\usage{
city
}
\description{
A shapefile containing the City of Takoma Park's boundaries
}
\keyword{datasets}
