% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_standardize.R
\name{label_html}
\alias{label_html}
\title{Turn Labels into HTML
A function to convert a character vector of labels for a leaflet map into a list with HTML applied to each element}
\usage{
label_html(label_vec)
}
\arguments{
\item{label_vec}{A character vector of labels corresponding to the length of the mapped dataframe}
}
\value{
A list with HTML applied to each element of the character vector
}
\description{
Turn Labels into HTML
A function to convert a character vector of labels for a leaflet map into a list with HTML applied to each element
}
\examples{
labs <- label_standardize(leafletwrappers::wards, "Ward name: {WARD}<p></p> Area: {Area}")

map <- leaflet::leaflet(leafletwrappers::wards)
leaflet::addPolygons(map, label = label_html(labs))
}
