% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_standardize.R
\name{label_standardize}
\alias{label_standardize}
\title{Create Leaflet Label}
\usage{
label_standardize(df, label_text)
}
\arguments{
\item{df}{A spatial dataframe that will be supplied to a leaflet() function}

\item{label_text}{A data-masked text string fed to a glue function and later read as HTML, so HTML elements should be incorporated (e.g., <p></p>)}
}
\value{
A character vector of labels that will be supplied in a leaflet map
}
\description{
Creates a label-vector for Leaflet map from a supplied dataframe and character string supplied to a glue function
}
\examples{
label_standardize(leafletwrappers::wards, "Ward name: {WARD}<p></p> Area: {Area}")
}
