% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpoly_legend.R
\name{layercontrolsquick}
\alias{layercontrolsquick}
\title{Quick layer controls}
\usage{
layercontrolsquick(leafletobj, groups, hide = NULL, .pos = "topleft")
}
\arguments{
\item{leafletobj}{Leaflet object}

\item{groups}{Character vector of overlay groups to show on Leaflet map}

\item{hide}{Groups to hide on leaflet map. If NULL (default), hides all groups but first group}

\item{.pos}{Position of layer controls; default "topleft"}
}
\value{
Leaflet map with overlay groups
}
\description{
Quick function to add a vector of groups as overlay groups for a leaflet map, and hide groups you don't want displayed in one step
}
