% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_numeric.R
\name{pal_numeric}
\alias{pal_numeric}
\title{Numeric palette}
\usage{
pal_numeric(var, colors, df, reverse = FALSE)
}
\arguments{
\item{var}{A variable in the dataframe, supplied as a string}

\item{colors}{Colors to supply to the palette argument in a colorNumeric() function}

\item{df}{A spatial dataframe with a stripped geography}

\item{reverse}{Whether to reverse the generated color palette}
}
\value{

}
\description{
Function to generate a color palette for a variable based on a supplied dataframe
}
