% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_pctile.R
\name{pal_pctile}
\alias{pal_pctile}
\title{Percentile palette
Function to generate a palette function to color percentile data}
\usage{
pal_pctile(colors = "Blues")
}
\arguments{
\item{colors}{Colors argument to leaflet::colorBin() function}
}
\value{

}
\description{
Percentile palette
Function to generate a palette function to color percentile data
}
