% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wards}
\alias{wards}
\title{Ward boundaries for 2013-2022}
\format{
A spatial dataframe with 6 rows and 8 variables
\describe{
\item{OBJECTID}{Unique identifier for Ward}
\item{ID}{ID for Ward}
\item{WARD}{Ward}
\item{Area}{Area of Ward}
\item{Area_acre}{Area of Ward in acres}
\item{Shape_Leng}{Length of polygon}
\item{Shape_Area}{Area of Polygon}
\item{geometry}{Geometry of Ward}
}
}
\source{
{City staff}
}
\usage{
wards
}
\description{
A shapefile containing polygons for Ward boundaries from 2013-2022
}
\keyword{datasets}
