% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wards_new}
\alias{wards_new}
\title{Ward boundaries for 2022-2030}
\format{
A spatial dataframe with 6 rows and 8 variables
\describe{
\item{DISTRICT}{Ward}
\item{DISTNUM}{Ward number}
\item{WARD}{Ward}
\item{Shape_Length}{Length of polygon}
\item{Shape_Area}{Area of Polygon}
\item{geometry}{Geometry of Ward}
}
}
\source{
{Redistricting contractor}
}
\usage{
wards_new
}
\description{
A shapefile containing Ward boundaries from 2022-2030 after the 2020 redistricting
}
\keyword{datasets}
