% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addlegend_standard.R
\name{addlegend_standard}
\alias{addlegend_standard}
\title{Add leaflet legend
A function to add a legend in standard format to a leaflet map}
\usage{
addlegend_standard(basemap, df, pal_funct, variable, group, title)
}
\arguments{
\item{basemap}{A leaflet basemap}

\item{df}{A spatial dataframe with the geography stripped that provides data visualized in the polygon}

\item{pal_funct}{A color function; e.g. leafletwrappers::color_numeric will often be approriate for numeric functions}

\item{variable}{The variable to create a legend for from the provided dataframe}

\item{group}{The group of the legend}

\item{title}{The title of the legend}
}
\value{
A legend in a Leaflet map
}
\description{
Add leaflet legend
A function to add a legend in standard format to a leaflet map
}
