% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpoly_legend.R
\name{addpoly_standard}
\alias{addpoly_standard}
\title{Add Polygons
A function to add a filled-in polygon layer coloring the polygon by the provided variable}
\usage{
addpoly_standard(
  basemap,
  df,
  pal_funct,
  variable,
  group,
  labels,
  .weight = 1,
  .color = "transparent",
  .opacity = 1,
  .fillopacity = 0.4,
  .data = leaflet::getMapData(basemap),
  .label_padding = "0.2px 0.2px",
  .label_lineheight = 0.8,
  .label_textsize = "10px",
  .label_opacity = 0.8
)
}
\arguments{
\item{basemap}{A leaflet basemap}

\item{df}{A spatial dataframe with the geography stripped that provides data visualized in the polygon}

\item{pal_funct}{A color function; e.g. leafletwrappers::color_numeric will often be approriate for numeric functions}

\item{variable}{The variable to visualize in the polygon from the provided dataframe}

\item{group}{The group-name of the polygon; a string}

\item{labels}{A list of labels corresponding to the length of the dataframe to appear as pop-ups}

\item{.weight}{The weight of the polygon lines. Default value of 1}

\item{.color}{The opacity of the polygon lines. Default value of 1}

\item{.opacity}{The opacity of the colors inside the polygons. Default value of 0.4}

\item{.fillopacity}{The color of the polygon lines. Default value of transparent}

\item{.data}{A spatial dataframe. Defaults to the basemap dataframe; if the data being visualized is different from the dataframe used in the basemap, supply the other dataframe here}

\item{.label_padding}{Label padding of label; default}

\item{.label_lineheight}{Label line-height; default 0.8}

\item{.label_textsize}{Label textsize; default "10px"}

\item{.label_opacity}{Label opacity; default 0.8}
}
\value{
A polygon layer in a leaflet map
}
\description{
Add Polygons
A function to add a filled-in polygon layer coloring the polygon by the provided variable
}
