% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_standardize.R
\name{label_output}
\alias{label_output}
\title{Label Output
Wrapper for label_standardize and label_html, to create labels for a leaflet map in one step}
\usage{
label_output(df, label_text)
}
\arguments{
\item{df}{Dataframe to create labels from}

\item{label_text}{Text fed to glue function to create labels using columns in dataframe}
}
\value{
List of strings corresponding to length of dataframe, supplied to leaflet function
}
\description{
Label Output
Wrapper for label_standardize and label_html, to create labels for a leaflet map in one step
}
\examples{
labs <- "Ward name: {WARD}<p></p> Area: {Area}"

map <- leaflet::leaflet(leafletwrappers::wards)
leaflet::addPolygons(map, label = label_output(leafletwrappers::wards, labs))
}
